Update renamed .d.ts files to work as .ts
