+++ packages/ilmomasiina-frontend/src/modules/editSignup/types.ts
@@ -1,16 +1,6 @@
 import { Signup } from '@tietokilta/ilmomasiina-models/src/services/signups';
-import {
-  resetState,
-  signupDeleted,
-  signupDeleteFailed,
-  signupLoaded,
-  signupLoadFailed,
-  signupSubmitting,
-  signupUpdated,
-  signupUpdateFailed,
-} from './actions';
 
-interface EditSignupState {
+export interface EditSignupState {
   event: Signup.Details.Event | null;
   signup: Signup.Details.Signup | null;
   loadError: boolean;
@@ -19,12 +9,4 @@ interface EditSignupState {
   deleted: boolean;
 }
 
-type EditSignupActions =
-  | ReturnType<typeof signupLoaded>
-  | ReturnType<typeof signupLoadFailed>
-  | ReturnType<typeof signupSubmitting>
-  | ReturnType<typeof signupUpdateFailed>
-  | ReturnType<typeof signupUpdated>
-  | ReturnType<typeof signupDeleted>
-  | ReturnType<typeof signupDeleteFailed>
-  | ReturnType<typeof resetState>;
+export type { EditSignupActions } from './actions';
