+++ packages/ilmomasiina-frontend/src/modules/events/types.ts
@@ -1,12 +1,8 @@
 import { Event } from '@tietokilta/ilmomasiina-models/src/services/events';
-import { eventsLoaded, eventsLoadFailed, resetState } from './actions';
 
-interface EventsState {
+export interface EventsState {
   events: Event.List | null;
   eventsLoadError: boolean;
 }
 
-type EventsActions =
-  | ReturnType<typeof eventsLoaded>
-  | ReturnType<typeof eventsLoadFailed>
-  | ReturnType<typeof resetState>;
+export type { EventsActions } from './actions';
