+++ packages/ilmomasiina-frontend/src/store/types.ts
@@ -1,4 +1,4 @@
-import { Action, AnyAction } from 'redux';
+import { Action, AnyAction, Store } from 'redux';
 import { ThunkAction, ThunkDispatch } from 'redux-thunk';
 
 import { AdminActions } from '../modules/admin/types';
@@ -8,29 +8,29 @@ import { EventsActions } from '../modules/events/types';
 import { SingleEventActions } from '../modules/singleEvent/types';
 import { makeRootReducer } from './reducers';
 
-type AppReducer = ReturnType<typeof makeRootReducer>;
-type AppState = ReturnType<AppReducer>;
+export type AppReducer = ReturnType<typeof makeRootReducer>;
+export type AppState = ReturnType<AppReducer>;
 
-type AppActions =
+export type AppActions =
   | AdminActions
   | EditorActions
   | EditSignupActions
   | EventsActions
   | SingleEventActions;
 
-type AsyncAction<R = void> = ThunkAction<
+export type AsyncAction<R = void> = ThunkAction<
 Promise<R>,
 AppState,
 undefined,
 AnyAction
 >;
 
-type DispatchAction<T extends AnyAction = Action> = ThunkDispatch<
+export type DispatchAction<T extends AnyAction = Action> = ThunkDispatch<
 AppState,
 undefined,
 T
 >;
 
-type GetState = () => AppState;
+export type GetState = () => AppState;
 
-type AppStore = Store<AppState, AppActions>;
+export type AppStore = Store<AppState, AppActions>;
